Add files via upload
This Python program is designed to facilitate the conversion of numbers between different numerical bases, namely: binary, decimal, octal, and hexadecimal. It provides a user-friendly interface through a series of menus, allowing users to input numbers in one base and convert them to another base of their choice.
How to Use:
Insert a Number: Choose option 'a' in Menu 1 to insert a new number for conversion.
Select Number Base: In Menu 2, choose the base of the number you're inputting (binary, decimal, octal, or hexadecimal).
Choose Conversion: After entering the number and base, Menu 3 prompts you to select the target base for conversion.
View Result: The program will perform the conversion and display the result.